requirements.txt
openai-whisper == 20231117
ffmpeg